% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_boxes.R
\name{add_row_boxes}
\alias{add_row_boxes}
\title{add_row_boxes
Add boxes around rows in a facet grid plot}
\usage{
add_row_boxes(
  p,
  rows_to_box = NULL,
  box_color = "black",
  box_linewidth = 1,
  buffer = 0.5
)
}
\arguments{
\item{p}{A ggplot object.}

\item{rows_to_box}{A character vector with the row facet variables to box. If NULL, all rows will be boxed.}

\item{box_color}{The color of the box. Default is "black".}

\item{box_linewidth}{The width of the box line. Default is 1.}

\item{buffer}{The buffer space between the box and the data points. Default is 0.5.}
}
\value{
A list of ggplot layers to add to the plot.
}
\description{
add_row_boxes
Add boxes around rows in a facet grid plot
}
