% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{mBP_col_labels}
\alias{mBP_col_labels}
\alias{mBP_row_labels}
\title{mBP_col_labels
Add matrix barplot labels to a plot}
\usage{
mBP_col_labels(g, margin = 3)

mBP_row_labels(g, margin = 3)
}
\arguments{
\item{g}{A ggplot object as returned from `matrix_barplot`.}

\item{margin}{A numeric value to adjust the width of the margin}
}
\value{
A ggplot object with labels added.
}
\description{
mBP_col_labels
Add matrix barplot labels to a plot

mBP_row_labels
Add matrix barplot row labels to a plot
}
