% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{mBP_labels}
\alias{mBP_labels}
\title{mBP_labels
Add matrix barplot labels to a plot}
\usage{
mBP_labels(g)
}
\arguments{
\item{g}{A ggplot object as returned from `matrix_barplot`.}
}
\value{
A ggplot object with labels added.
}
\description{
mBP_labels
Add matrix barplot labels to a plot
}
