% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{mBP_legend}
\alias{mBP_legend}
\title{mBP_legend
Add matrix barplot legend to a plot}
\usage{
mBP_legend(g, legend.position = "bottom", ...)
}
\arguments{
\item{g}{A ggplot object as returned from `matrix_barplot`.}

\item{legend.position}{A character string indicating the position of the legend. Default is "bottom".}

\item{...}{Parameters passed to `patchwork::plot_layout` for defining the plot layout. See ?plot_layout for more information.}
}
\value{
A ggplot object with a legend added.
}
\description{
mBP_legend
Add matrix barplot legend to a plot
}
\note{
This function will throw a warning about removing rows containing missing values or values outside the scale range due to some hidden features that are not plotted.
}
