% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_barplot.R
\name{stat_centerBar}
\alias{stat_centerBar}
\title{stat_centerBar
A new stat for ggplot2 that creates a barplot or line segments with bars centered around the x-axis}
\usage{
stat_centerBar(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by `aes`.}

\item{data}{The data to be displayed in this layer.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlappling points on this layer}

\item{na.rm}{A logical indicating whether NA values should be removed from the data before plotting}

\item{show.legend}{A logical indicating whether this layer should be included in the legend}

\item{inherit.aes}{A logical indicating whether to inherit the default aesthetics}

\item{...}{Other arguments passed on to `layer`}
}
\value{
A ggproto object
}
\description{
stat_centerBar
A new stat for ggplot2 that creates a barplot or line segments with bars centered around the x-axis
}
\details{
For specifics on creating a new stat, see [this documentation](https://bookdown.org/rdpeng/RProgDA/building-new-graphical-elements.html#example-normal-confidence-intervals)
}
