% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StatCenterBar}
\alias{StatCenterBar}
\title{StatCenterBar}
\format{
An object of class \code{StatCenterBar} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatCenterBar
}
\description{
A new stat for ggplot2 that creates a barplot or line segments with bars centered around the x-axis
}
\keyword{datasets}
