% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_barplot.R
\name{matrix_barplot}
\alias{matrix_barplot}
\title{matrix_barplot
Creation of a matrix barplot using ggplot2}
\usage{
matrix_barplot(
  data = NULL,
  mapping = NULL,
  rows = vars(),
  cols = vars(),
  facet_colors = deeptime::stages,
  orientation = "landscape",
  switch = FALSE,
  show_strip.x.text = FALSE,
  show_strip.y.text = FALSE,
  show_strip.y.background = orientation == "portrait",
  show_strip.x.background = orientation == "landscape",
  remove.x.labels = TRUE,
  remove.y.labels = TRUE,
  remove.panel.grid = TRUE,
  ...
)
}
\arguments{
\item{data}{The data to be displayed in this layer.}

\item{mapping}{Set of aesthetic mappings created by `aes`.}

\item{rows}{A set of aesthetic mappings for the rows of the matrix barplot. Rows will be sorted alphabetically for character variables and numerically for factors.}

\item{cols}{A set of aesthetic mappings for the columns of the matrix barplot. Columns will be sorted alphabetically for character variables and numerically for factors.}

\item{facet_colors}{Specifies which colors to use to replace the strip backgrounds (i.e. corresponding to the values in the variables specified in `rows` and `cols`. Either A) a function that returns a color for a given strip label, B) the character name of a function that does the same, C) a named character vector with names matching strip labels and values indicating the desired colors, or D) a data.frame representing a lookup table with columns named "name" (matching strip labels) and "color" (indicating desired colors).}

\item{orientation}{Character value defining the orientation of the plot. Can be "landscape" or "portrait".}

\item{switch}{Logical. By default (`switch = FALSE`), the labels are displayed on the bottom and/or left of the plot. If `switch = TRUE`, the labels are displayed on the top and/or right of the plot.}

\item{show_strip.x.text}{Logical indicating whether to show the x-axis facet label text inside the colored rectangles associated with `cols`.}

\item{show_strip.y.text}{Logical indicating whether to show the y-axis facet label text inside the colored rectangles associated with `rows`.}

\item{show_strip.y.background}{Logical indicating whether to print rectangles for the y-axis facets associated with `rows`.}

\item{show_strip.x.background}{Logical indicating whether to print rectangles for the x-axis facets associated with `cols`.}

\item{remove.x.labels}{Logical indicating whether to remove the x-axis labels.}

\item{remove.y.labels}{Logical indicating whether to remove the y-axis labels.}

\item{remove.panel.grid}{Logical indicating whether to remove the panel grid.}

\item{...}{Other arguments passed on to `ggplot`. These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3.}
}
\value{
A ggplot2 object
}
\description{
matrix_barplot
Creation of a matrix barplot using ggplot2
}
\details{
This function creates a matrix barplot and expects `x` (e.g. a gene or metabolite) and `y` (a numeric value defining the hight of the bars, e.g. -log10 p-value) to be defined in the mapping.
The colors of the bars are defined by the `color` aesthetic.

Fill colors for the categories along the x-axis unless there are more than 12 categories.
In this case, or when category labels are desired in lieu of colors, the `label` aesthetic can be used.
}
