% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_barplot.R
\name{matrix_barplot}
\alias{matrix_barplot}
\title{matrix_barplot
Creation of a matrix barplot using ggplot2}
\usage{
matrix_barplot(
  data = NULL,
  mapping = NULL,
  rows = vars(),
  cols = vars(),
  facet_colors = deeptime::stages,
  orientation = "landscape",
  switch = FALSE,
  pval_cap = 1e-05,
  legend.position = "top",
  sig_lines_at = NULL,
  sig_line_type = NULL,
  sig_line_labels = NULL,
  sig_line_color = "grey80",
  sig_line_alpha = NULL,
  show_strip.x.text = FALSE,
  show_strip.y.text = FALSE,
  show_strip.y.background = orientation == "portrait",
  show_strip.x.background = orientation == "landscape",
  remove.x.labels = TRUE,
  remove.y.labels = TRUE,
  remove.panel.grid = TRUE,
  ...
)
}
\arguments{
\item{data}{The data to be displayed in this layer.}

\item{mapping}{Set of aesthetic mappings created by `aes`.}

\item{rows}{A set of aesthetic mappings for the rows of the matrix barplot. Rows will be sorted alphabetically for character variables and numerically for factors.}

\item{cols}{A set of aesthetic mappings for the columns of the matrix barplot. Columns will be sorted alphabetically for character variables and numerically for factors.}

\item{facet_colors}{Specifies which colors to use to replace the strip backgrounds (i.e. corresponding to the values in the variables specified in `rows` and `cols`. Either A) a function that returns a color for a given strip label, B) the character name of a function that does the same, C) a named character vector with names matching strip labels and values indicating the desired colors, or D) a data.frame representing a lookup table with columns named "name" (matching strip labels) and "color" (indicating desired colors).}

\item{orientation}{Character value defining the orientation of the plot. Can be "landscape" or "portrait".}

\item{switch}{Logical. By default (`switch = FALSE`), the labels are displayed on the bottom and/or left of the plot. If `switch = TRUE`, the labels are displayed on the top and/or right of the plot.}

\item{pval_cap}{Numeric value indicating the maximum p-value to plot. All p-values above this value will be plotted at this value with an asterisk indicating they are capped.}

\item{legend.position}{Character value indicating the position of the legend. Can be "top", "bottom", "left", "right", or "none".}

\item{sig_lines_at}{Numeric vector indicating the significance levels (i.e. p-values) to mark on the plot.}

\item{sig_line_type}{Character value indicating the type of line to use for the significance lines (should be the same length as `sig_lines_at` or a multiple). Can be "solid", "dashed", "dotted", "dotdash", "longdash", or "twodash".}

\item{sig_line_labels}{Character vector indicating the labels to use for the significance lines.}

\item{sig_line_color}{Character value indicating the color to use for the significance lines.}

\item{sig_line_alpha}{Numeric value indicating the alpha transparency to use for the significance lines.}

\item{show_strip.x.text}{Logical indicating whether to show the x-axis facet label text inside the colored rectangles associated with `cols`.}

\item{show_strip.y.text}{Logical indicating whether to show the y-axis facet label text inside the colored rectangles associated with `rows`.}

\item{show_strip.y.background}{Logical indicating whether to print rectangles for the y-axis facets associated with `rows`.}

\item{show_strip.x.background}{Logical indicating whether to print rectangles for the x-axis facets associated with `cols`.}

\item{remove.x.labels}{Logical indicating whether to remove the x-axis labels.}

\item{remove.y.labels}{Logical indicating whether to remove the y-axis labels.}

\item{remove.panel.grid}{Logical indicating whether to remove the panel grid.}

\item{...}{Other arguments passed on to `ggplot`. These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3.}
}
\value{
A ggplot2 object
}
\description{
matrix_barplot
Creation of a matrix barplot using ggplot2
}
\details{
This function creates a matrix barplot and expects `x` (e.g. a gene or metabolite) and `y` (a numeric value defining the height of the bars, e.g. -log10 p-value) to be defined in the mapping.
Other parameter defaults are tuned to work well with p-values (y-axis) and fold change statistics (mapped to the color aesthetic).
The colors of the bars are defined by the `color` aesthetic.

Fill colors for the categories along the x-axis unless there are more than 12 categories.
In this case, or when category labels are desired in lieu of colors, the `label` aesthetic can be used.

When `sig_lines_*` are `NULL`, `matrix_barplot` will attempt to guess appropriate significance levels, line attributes based on the minimum value of `y`.
If the length of `sig_line_` `type`, `labels` and `color` are not equal to the length of `sig_lines_at`, `matrix_barplot` will recycle the values provided.
To disable plotting of significance lines, set `sig_lines_at = NA`. Note that `sig_line_color` must be a single value.
}
